feat: remove policies and audit for MVP simplification
Remove the Policy model and all related UI/actions. Secrets now match
directly based on host pattern for any agent belonging to the user,
without the extra policy layer connecting agents to secrets.

Remove the AuditLog model, audit page, audit API endpoint, and all
audit-related UI including the recent audit table and stats on the
overview page. The overview now shows only agent and secret counts.
