+++ .env.example
@@ -13,6 +13,6 @@ GOOGLE_CLIENT_SECRET=
 
 
 # Secret encryption key for encrypting stored secrets (API keys, tokens, etc.)
-# Required when using the proxy with secrets/policies.
+# Required when using the proxy with secrets.
 # Generate with: node -e "console.log(require('crypto').randomBytes(32).toString('base64'))"
 SECRET_ENCRYPTION_KEY=
