+++ apps/web/src/app/(dashboard)/overview/_components/overview-content.tsx
@@ -2,38 +2,15 @@
 
 import { useEffect, useState } from "react";
 import { useAuth } from "@/providers/auth-provider";
-import {
-  getRecentAuditLogs,
-  getAuditStats,
-  getProxyCounts,
-} from "@/lib/actions/audit";
+import { getProxyCounts } from "@/lib/actions/counts";
 import { ApiKeyCard } from "./api-key-card";
 import { StatsCards } from "./stats-cards";
-import { RecentAuditTable } from "./recent-audit-table";
-
-interface AuditLogEntry {
-  id: string;
-  action: string;
-  service: string;
-  status: string;
-  source: string;
-  createdAt: Date;
-}
-
-interface AuditStats {
-  totalActions: number;
-  recentActions: number;
-  serviceCount: number;
-}
 
 export function OverviewContent() {
   const { user } = useAuth();
-  const [logs, setLogs] = useState<AuditLogEntry[]>([]);
-  const [stats, setStats] = useState<AuditStats | null>(null);
   const [proxyCounts, setProxyCounts] = useState({
     agents: 0,
     secrets: 0,
-    policies: 0,
   });
   const [loading, setLoading] = useState(true);
 
@@ -41,13 +18,7 @@ export function OverviewContent() {
     if (!user?.id) return;
 
     setLoading(true);
-    Promise.all([
-      getRecentAuditLogs(5, user.id),
-      getAuditStats(user.id),
-      getProxyCounts(user.id),
-    ]).then(([logsData, statsData, proxy]) => {
-      setLogs(logsData);
-      setStats(statsData);
+    getProxyCounts(user.id).then((proxy) => {
       setProxyCounts(proxy);
       setLoading(false);
     });
@@ -65,16 +36,10 @@ export function OverviewContent() {
       <ApiKeyCard />
 
       <StatsCards
-        totalActions={stats?.totalActions ?? 0}
-        recentActions={stats?.recentActions ?? 0}
-        serviceCount={stats?.serviceCount ?? 0}
         agentCount={proxyCounts.agents}
         secretCount={proxyCounts.secrets}
-        policyCount={proxyCounts.policies}
         loading={loading}
       />
-
-      <RecentAuditTable logs={logs} />
     </div>
   );
 }
